requirements.txt
attrs==19.1.0
beautifulsoup4==4.8.0
blis==0.2.4
Bottleneck==1.2.1
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
cymem==2.0.2
Django==2.2.4
djangorestframework==3.10.2
fastprogress==0.1.21
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
murmurhash==1.0.2
numexpr==2.6.9
numpy==1.17.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
packaging==19.1
pandas==0.25.0
Pillow==6.1.0
pkg-resources==0.0.0
plac==0.9.6
preshed==2.0.1
pyparsing==2.4.2
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
scipy==1.3.0
six==1.12.0
soupsieve==1.9.2
spacy==2.1.8
sqlparse==0.3.0
srsly==0.0.7
thinc==7.0.8
tqdm==4.32.2
tweepy==3.8.0
urllib3==1.25.3
wasabi==0.2.2
